% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_draw.R
\name{interpro2draw}
\alias{interpro2draw}
\title{Title}
\usage{
interpro2draw(
  interpro,
  gene_list,
  gene_length,
  domain_db = "Pfam",
  motif_db = "TMHMM"
)
}
\arguments{
\item{motif_db}{}
}
\value{

}
\description{
Title
}
