% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_seq.R
\name{rename_seq}
\alias{rename_seq}
\title{Convert gff to gggenomes data}
\usage{
rename_seq(gff_data, gene_list, aten = TRUE)
}
\arguments{
\item{gff_data}{gff tible from gff files using read_gff function}

\item{gene_list}{genelist provided by blast_extract or domain_extract}

\item{aten}{if your gff data is acropora tenuis genome (default = TRUE)}
}
\value{
gff tibble coressponding to specified gene_list
}
\description{
Convert gff to gggenomes data
}
\examples{

}
