% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{blast_extract}
\alias{blast_extract}
\title{Extract the reference genes corresponding to the query genes from the blast results}
\usage{
blast_extract(blast_data, bacteria_gene, evalues = 1 * 10^-10)
}
\arguments{
\item{evalues}{}
}
\value{

}
\description{
Extract the reference genes corresponding to the query genes from the blast results
}
