% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{domain_extract}
\alias{domain_extract}
\title{Extract the genes corresponding to the domains from the inteproscan results}
\usage{
domain_extract(interpro, domain)
}
\arguments{
\item{domain}{}
}
\value{

}
\description{
Extract the genes corresponding to the domains from the inteproscan results
}
