% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_seq.R
\name{rename_seq}
\alias{rename_seq}
\title{Convert gff to gggenomes data}
\usage{
rename_seq(gff_data, gene_list, aten = TRUE)
}
\arguments{
\item{aten}{}
}
\value{

}
\description{
Convert gff to gggenomes data
}
